requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
ImageHash==4.3.2
numpy==2.0.2
pillow==11.1.0
PyWavelets==1.6.0
requests==2.32.3
scipy==1.13.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
